Create a sprite sheet for Player.
Changed player sprite sheet to one of the sprites from the dungeon tilsets. Removed sword from player sprite and instead created a way for the player to "Equip a sword instead". Added hearts meter for Player. In the future we may remove the health bra completely and just do a hit/heart meter to check if player has died.